+++ src/modules/Lightbox/index.js
@@ -140,6 +140,7 @@ export default class Lightbox {
 
     this.lightboxes.forEach(lightbox => {
       const href = lightbox.getAttribute('data-lightbox')
+      const srcset = lightbox.getAttribute('data-srcset')
       const originalImage = lightbox.querySelector('img')
       const alt = originalImage.getAttribute('alt')
       const section = lightbox.getAttribute('data-lightbox-section') || 'general'
@@ -154,7 +155,8 @@ export default class Lightbox {
 
       const image = {
         href,
-        alt
+        alt,
+        srcset
       }
 
       const index = this.sections[section].push(image) - 1
@@ -337,6 +339,7 @@ export default class Lightbox {
     for (let x = 0; x < 3; x += 1) {
       if (this.imgs[index + x]) {
         this.imgs[index + x].src = this.sections[section][index + x].href
+        this.imgs[index + x].setAttribute('srcset', this.sections[section][index + x].srcset)
       } else {
         break
       }
@@ -344,6 +347,7 @@ export default class Lightbox {
 
     this.nextImage = this.imgs[index]
     this.nextImage.src = this.sections[section][index].href
+    this.nextImage.setAttribute('srcset', this.sections[section][index].srcset)
 
     this.opts.onImageIn(this)
     this.timelines.image.call(() => {